teams/WYM/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
motor==3.7.1
pymongo==4.12.1
pydantic==2.11.3
pydantic-settings==2.9.1
httpx==0.28.1
python-dotenv==1.1.0
groq==0.25.0
